fix: adapt interface for incomplete dbus reply messages
The ModemManager dbus interface is not always reliable in its return values.
This makes parsing the replies difficult. Previously, we assumed that successful
API calls always result in docs conforming replies. This, however, is not the
case. With this commit, we instead return the APIs result as is (that is: a
HashMap) for the caller to evaluate on their own. We still provide concrete
modemmanager types as they come in handy later.