fix: change interface for robustness on incomplete modemmanager replies (#6)
* fix: correct spelling of bearer properties field name

One field type of the bearer properties was spelled wrong, which caused the
bearer properties not to be reported correctly. With this commit we fix the
spelling of the field name.

* build: introduce conventional-commits

We switch to conventional commits format for the commit messages. While doing
so, we additionally add a github action to check for proper commit format and
remove the CHANGELOG, as the latter can automatically be generated from the
commits.

* feat: add type aliases for ip configuration types

We extend the crate with type aliases for the IPv4 and IPv6 two specializations.

* fix: adapt interface for incomplete dbus reply messages

The ModemManager dbus interface is not always reliable in its return values.
This makes parsing the replies difficult. Previously, we assumed that successful
API calls always result in docs conforming replies. This, however, is not the
case. With this commit, we instead return the APIs result as is (that is: a
HashMap) for the caller to evaluate on their own. We still provide concrete
modemmanager types as they come in handy later.

* fix: adjust to correct version 0.3.0